Revert "fixup! modules: sql: afk: type"
This reverts commit 5355edd9fa0d086057b001f426008598060d17c3.

+++ tg_bot/modules/sql/afk_sql.py
@@ -11,7 +11,7 @@ class AFK(BASE):
     user_id = Column(BigInteger, primary_key=True)
     is_afk = Column(Boolean)
     reason = Column(UnicodeText)
-    afk_time = Column(BigInteger)
+    afk_time = Column(Integer)
 
     def __init__(self, user_id, afk_time, reason="", is_afk=True):
         self.user_id = user_id
